boke/requirements.txt
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Markdown==3.8
bleach==6.2.0
